refs:40721 Added helper and fixed spec for Consumerrole
